// frontend/src/app/models/settings/user-info.model.ts
export interface UserInfo {
    userId: number;
	emailAddress: string;
	phoneNumber: string;
	firstName: string;
	lastName: string;
	displayName: string;
	zipCode: string;
	userAddress: string;
	userCountry: string;
	userCity: string;
	userState: string;
	introText: string;
	birthday: Date;
	pronouns: string;
	sexualOrientation: string;
	religion: string;
	politicalAffiliation: string;
	ethnicity: string;
	ethnicitySub: string;
	profilePicture: string;
	lastAccessedDate: Date;
	isActive: boolean;
	isVerifiedEmail: boolean;
	isVerifiedIdentity: boolean;
	insertDate: Date;
	updateDate: Date;
}